communication_engineering_serialport_helper/requirements.txt
altgraph==0.16.1
apipkg==1.5
atomicwrites==1.2.1
attrs==18.2.0
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.0
concurrencytest==0.1.2
Django==2.1.7
et-xmlfile==1.0.1
execnet==1.5.0
extras==1.0.0
fixtures==3.0.0
future==0.17.1
get==2018.11.19
html-testRunner==1.1.2
idna==2.8
importlib-resources==1.0.2
iso8601==0.1.12
jdcal==1.4
Jinja2==2.9.5
linecache2==1.0.0
lml==0.0.6
lxml==4.2.5
macholib==1.11
MarkupSafe==1.1.0
MonthDelta==0.9.1
more-itertools==4.3.0
openpyxl==2.5.9
pbr==5.1.1
pefile==2018.8.8
pluggy==0.8.0
post==2018.11.20
public==2018.11.20
py==1.7.0
pyexcel==0.5.9.1
pyexcel-io==0.5.9.1
pyexcel-xls==0.5.8
pyexcel-xlsx==0.5.6
pyexcel-xlsxr==0.5.2
pyexcel-xlsxw==0.4.2
pyexcel-xlsxwx==0.1.7
PyInstaller==3.4
pymssql==2.1.4
pypiwin32==223
pyserial==3.4
pytest==3.9.3
pytest-assume==1.2
pytest-forked==0.2
pytest-xdist==1.24.1
python-dateutil==2.7.5
python-mimeparse==1.6.0
python-subunit==1.3.0
pytz==2018.9
pywin32==224
pywin32-ctypes==0.2.0
PyYAML==3.13
query-string==2018.11.20
request==2018.11.20
requests==2.21.0
ruamel.yaml==0.15.76
selenium==3.14.1
six==1.11.0
testtools==2.3.0
texttable==1.5.0
traceback2==1.4.0
unittest2==1.1.0
urllib3==1.24
xlrd==1.1.0
XlsxWriter==1.1.2
xlwt==1.3.0
xmlrunner==1.7.7